Module AltErgoLib.Arith

val calc_power : Numbers.Q.t -> Numbers.Q.t -> Ty.t -> Numbers.Q.t

calc_power x y t Compute x^y. Raise Exit if y is not an Int (castable in Int).

val calc_power_opt : Numbers.Q.t -> Numbers.Q.t -> Ty.t -> Numbers.Q.t option

Same as calc_power but return an option. Return None if the exception Exit is raised by calc_power

module Type : functor (X : Sig.X) -> Polynome.T with type r = X.r

module Shostak : functor (X : Sig.X) -> functor (P : Polynome.EXTENDED_Polynome with type r = X.r) -> Sig.SHOSTAK with type r = X.r and type t = P.t