Module AltErgoLib.Fpa_rounding

type rounding_mode =

| NearestTiesToEven

| ToZero

| Up

| Down

| NearestTiesToAway

| Aw

| Od

| No

| Nz

| Nd

| Nu

val fpa_rounding_mode : Ty.t

val rounding_mode_of_hs : Hstring.t -> rounding_mode

val integer_log_2 : Numbers.Q.t -> int

Integer part of binary logarithm for NON-ZERO POSITIVE number *

val float_of_rational : int -> int -> rounding_mode -> Numbers.Q.t -> Numbers.Q.t * Numbers.Z.t * int

float_of_rational prec exp mode x float approx of a rational constant. The function also returns the mantissa and the exponent. i.e. if res, m, e = float_of_rational prec exp mode x, then res = m * 2^e *

val round_to_integer : rounding_mode -> Numbers.Q.t -> Numbers.Q.t

round_to_integer mode x rounds the rational x to an integer depending on the rounding mode mode

val empty_cache : unit -> unit

Empties the module's inner cache