Module type Hconsing.S

type t

The type of value used.

val make : t -> t

Hashcons a value t, either returning t, or a value equal to t that was hashconsed previously.

val elements : unit -> t list

Returns the list of all unique hashconsed elements.