Module AltErgoLib.Input

Input method

exception Method_not_registered of string

Exceptions raised when trying to lookup an input method that has not been registered.

module type S = sig ... end

This modules defines an input method. Input methods are responsible for two things: parsing and typechceking either an input file (possibly with some preludes files), or arbitrary terms. This last functionality is currently only used in the GUI.

val register : string -> (module S) -> unit

Register a new input method.

val find : string -> (module S)

Find an input method by name.

raises Method_not_registered

if the name is not registered.