Module AltErgoLib__Arith

val calc_power : AltErgoLib.Numbers.Q.t -> AltErgoLib.Numbers.Q.t -> AltErgoLib.Ty.t -> AltErgoLib.Numbers.Q.t

calc_power x y t Compute x^y. Raise Exit if y is not an Int (castable in Int).

val calc_power_opt : AltErgoLib.Numbers.Q.t -> AltErgoLib.Numbers.Q.t -> AltErgoLib.Ty.t -> AltErgoLib.Numbers.Q.t option

Same as calc_power but return an option. Return None if the exception Exit is raised by calc_power

module Type : functor (X : AltErgoLib.Sig.X) -> AltErgoLib.Polynome.T with type r = X.r

module Shostak : functor (X : AltErgoLib.Sig.X) -> functor (P : AltErgoLib.Polynome.EXTENDED_Polynome with type r = X.r) -> AltErgoLib.Sig.SHOSTAK with type r = X.r and type t = P.t