Module AltErgoLib__Errors

Errors module

Error types

type typing_error =

| BitvExtract of int * int

| BitvExtractRange of int * int

| ClashType of string

| ClashLabel of string * string

| ClashParam of string

| TypeDuplicateVar of string

| UnboundedVar of string

| UnknownType of string

| WrongArity of string * int

| SymbAlreadyDefined of string

| SymbUndefined of string

| NotAPropVar of string

| NotAPredicate of string

| Unification of AltErgoLib.Ty.t * AltErgoLib.Ty.t

| ShouldBeApply of string

| WrongNumberofArgs of string

| ShouldHaveType of AltErgoLib.Ty.t * AltErgoLib.Ty.t

| ShouldHaveTypeIntorReal of AltErgoLib.Ty.t

| ShouldHaveTypeInt of AltErgoLib.Ty.t

| ShouldHaveTypeBitv of AltErgoLib.Ty.t

| ArrayIndexShouldHaveTypeInt

| ShouldHaveTypeArray

| ShouldHaveTypeRecord of AltErgoLib.Ty.t

| ShouldBeARecord

| ShouldHaveLabel of string * string

| NoLabelInType of AltErgoLib.Hstring.t * AltErgoLib.Ty.t

| ShouldHaveTypeProp

| NoRecordType of AltErgoLib.Hstring.t

| DuplicateLabel of AltErgoLib.Hstring.t

| DuplicatePattern of string

| WrongLabel of AltErgoLib.Hstring.t * AltErgoLib.Ty.t

| WrongNumberOfLabels

| Notrigger

| CannotGeneralize

| SyntaxError

| ThExtError of string

| ThSemTriggerError

| WrongDeclInTheory

| ShouldBeADT of AltErgoLib.Ty.t

| MatchNotExhaustive of AltErgoLib.Hstring.t list

| MatchUnusedCases of AltErgoLib.Hstring.t list

| NotAdtConstr of string * AltErgoLib.Ty.t

| BadPopCommand of {

pushed : int;

to_pop : int;

}

| ShouldBePositive of int

Error that can be raised by the typechecker

type run_error =

| Invalid_steps_count of int

| Steps_limit of int

| Failed_check_unsat_core

| Unsupported_feature of string

| Dynlink_error of string

Errors that can be raised at solving

type error =

| Parser_error of string
Error used at parser loading

| Lexical_error of AltErgoLib.Loc.t * string
Error used by the lexer

| Syntax_error of AltErgoLib.Loc.t * string
Error used by the parser

| Typing_error of AltErgoLib.Loc.t * typing_error
Error used at typing

| Run_error of run_error
Error used during solving

| Warning_as_error

All types of error that can be raised

Exceptions

exception Error of error

Raising exceptions functions

val error : error -> 'a

Raise the input error as Error

val typing_error : typing_error -> AltErgoLib.Loc.t -> 'a

Raise the input typing_error as Typing_error

val run_error : run_error -> 'a

Raise the input run_error as Run_error

val warning_as_error : unit -> unit

Raise Warning_as_error as Error if the option warning-as-error is set This function can be use after warning

Printing

val report : Stdlib.Format.formatter -> error -> unit

Print a message on the formatter corresponding to the error