Module AltErgoLib__Hconsing.Make

Functor to create a hashconsing module from a module describing values.

Parameters

H : HASHED

Signature

type t = H.elt

The type of value used.

val make : t -> t

Hashcons a value t, either returning t, or a value equal to t that was hashconsed previously.

val elements : unit -> t list

Returns the list of all unique hashconsed elements.