Module AltErgoLib__Loc

Position in input files

This module defines a notion of location in files. Note: this only specifies a position in an arbitrary file, it does not contain information about the file itself.

type t = Stdlib.Lexing.position * Stdlib.Lexing.position

The type of locations, a location is made up of two position in the file, respectively corresponding to the beginning and end of the location.

val dummy : t

A dummy location.

val report : Stdlib.Format.formatter -> t -> unit

Report a location on the given formatter, using standard human-redable location reporting.