Module AltErgoLib__Typechecker

type env

The type of global environment of the typechecker.

val empty_env : env

The empty/initial environment

val type_expr : env -> (AltErgoLib.Symbols.t * AltErgoLib.Ty.t) list -> AltErgoLib.Parsed.lexpr -> int AltErgoLib.Typed.atterm

Typecheck an input expression (i.e. term (or formula ?)), given a local environment and a list of local types used to extend the initial environment.

raises Typing_error

Errors.Typing_error

val type_parsed : env -> env Stdlib.Stack.t -> AltErgoLib.Parsed.decl -> int AltErgoLib.Typed.atdecl list * env

Type a single declaration.

raises Typing_error

Errors.Typing_error

val type_file : AltErgoLib.Parsed.file -> (int AltErgoLib.Typed.atdecl * env) list * env

Type an input file. Returns the successive global environments obtained after typing each declaration.

raises Typing_error

Errors.Typing_error

val split_goals : (int AltErgoLib.Typed.atdecl * 'a) list -> ((int AltErgoLib.Typed.atdecl * 'a) list * string) list

val split_goals_and_cnf : (int AltErgoLib.Typed.atdecl * 'a) list -> (AltErgoLib.Commands.sat_tdecl list * string) list