Module AltErgoLib__Options

Options module used at start-up to parse the command line

val fmt : Format.formatter

Standard output formatter for Format

type model =

| MNone

| MDefault

| MAll

| MComplete

Type used to describe the type of models wanted

type input_format =

| Native
Native Alt-Ergo format

| Smtlib2
Smtlib default format

| Why3
Why3 file format

| Unknown of string
Unknown file format

Type used to describe the type of input wanted by set_input_format

type output_format = input_format

Type used to describe the type of output wanted by set_output_format

Setter functions

Setters for debug flags

val set_debug : bool -> unit

Set debug

val set_debug_ac : bool -> unit

Set debug_ac

val set_debug_adt : bool -> unit

Set debug_adt

val set_debug_arith : bool -> unit

Set debug_arith

val set_debug_arrays : bool -> unit

Set debug_arrays

val set_debug_bitv : bool -> unit

Set debug_bitv

val set_debug_cc : bool -> unit

Set debug_cc

val set_debug_combine : bool -> unit

Set debug_combine

val set_debug_constr : bool -> unit

Set debug_constr

val set_debug_explanations : bool -> unit

Set debug_explanations

val set_debug_fm : bool -> unit

Set debug_fm

val set_debug_gc : bool -> unit

Set debug_gc

val set_debug_ite : bool -> unit

Set debug_ite

val set_debug_matching : int -> unit

Set debug_matching

Possible values are

Disabled

Light

Full

val set_debug_sat : bool -> unit

Set debug_sat

val set_debug_sat_simple : bool -> unit

Set debug_sat_simple

val set_debug_split : bool -> unit

Set debug_split

val set_debug_sum : bool -> unit

Set debug_sum

val set_debug_types : bool -> unit

Set debug_types

val set_debug_typing : bool -> unit

Set debug_typing

val set_debug_uf : bool -> unit

Set debug_uf

val set_debug_unsat_core : bool -> unit

Set debug_unsat_core

val set_debug_use : bool -> unit

Set debug_use

val set_profiling : float -> bool -> unit

Set profiling

val set_timers : bool -> unit

Set timers

Additional setters

val set_type_only : bool -> unit

Set type_only

val set_age_bound : int -> unit

Set age_bound

val set_bottom_classes : bool -> unit

Set bottom_classes

val set_fm_cross_limit : AltErgoLib.Numbers.Q.t -> unit

Set fm_cross_limit

val set_frontend : string -> unit

Set frontend

val set_greedy : bool -> unit

Set greedy

val set_inline_lets : bool -> unit

Set inline_lets

val set_input_format : input_format -> unit

Set input_format

val set_interpretation : int -> unit

Set interpretation

Possible values are :

Unknown

Before instantiation

Before every decision or instantiation

A negative value (-1, -2, or -3) will disable interpretation display.

val set_max_split : AltErgoLib.Numbers.Q.t -> unit

Set max_split

val set_model : model -> unit

Set model

Possible values are :

Default

Complete

All

val set_nb_triggers : int -> unit

Set nb_triggers

val set_no_ac : bool -> unit

Set no_ac

val set_no_contracongru : bool -> unit

Set no_contragru

val set_no_ematching : bool -> unit

Set no_ematching

val set_no_nla : bool -> unit

Set no_nla

val set_no_user_triggers : bool -> unit

Set no_user_triggers

val set_normalize_instances : bool -> unit

Set normalize_instances

val set_output_format : output_format -> unit

Set output_format

val set_parse_only : bool -> unit

Set parse_only

val set_restricted : bool -> unit

Set restricted

val set_rewriting : bool -> unit

Set rewriting

val set_rules : int -> unit

Set rules

val set_save_used_context : bool -> unit

Set save_used_context

val set_steps_bound : int -> unit

Set steps_bound

val set_term_like_pp : bool -> unit

Set term_like_pp

val set_thread_yield : (unit -> unit) -> unit

Set thread_yield

val set_timelimit : float -> unit

Set timelimit

val set_timeout : (unit -> unit) -> unit

Set timeout

val set_triggers_var : bool -> unit

Set triggers_var

val set_type_smt2 : bool -> unit

Set type_smt2

val set_unsat_core : bool -> unit

Set unsat_core

val set_verbose : bool -> unit

Set verbose

val set_file_for_js : string -> unit

Updates the filename to be parsed and sets a js_mode flag

Getter functions

Getters for debug flags

val debug : unit -> bool

Get the debugging flag.

val debug_warnings : unit -> bool

Get the debugging flag of warnings.

val debug_cc : unit -> bool

Get the debugging flag of cc.

val debug_gc : unit -> bool

Prints some debug info about the GC's activity.

val debug_use : unit -> bool

Get the debugging flag of use.

val debug_uf : unit -> bool

Get the debugging flag of uf.

val debug_fm : unit -> bool

Get the debugging flag of inequalities.

val debug_fpa : unit -> int

Get the debugging value of floating-point.

Default to 0.

val debug_sum : unit -> bool

Get the debugging flag of Sum.

val debug_adt : unit -> bool

Get the debugging flag of ADTs.

val debug_arith : unit -> bool

Get the debugging flag of Arith (without fm).

val debug_bitv : unit -> bool

Get the debugging flag of bitv.

val debug_ac : unit -> bool

Get the debugging flag of ac.

val debug_sat : unit -> bool

Get the debugging flag of SAT.

val debug_sat_simple : unit -> bool

Get the debugging flag of SAT (simple output).

val debug_typing : unit -> bool

Get the debugging flag of typing.

val debug_constr : unit -> bool

Get the debugging flag of constructors.

val debug_arrays : unit -> bool

Get the debugging flag of arrays.

val debug_ite : unit -> bool

Get the debugging flag of ite.

val debug_types : unit -> bool

Get the debugging flag of types.

val debug_combine : unit -> bool

Get the debugging flag of combine.

val debug_unsat_core : unit -> bool

Replay unsat-cores produced by unsat_core. The option implies unsat_core is set to true.

val debug_split : unit -> bool

Get the debugging flag of case-split analysis.

val debug_matching : unit -> int

Get the debugging flag of E-matching

Possible values are

Disabled

Light

Full

val debug_explanations : unit -> bool

Get the debugging flag of explanations.

val debug_triggers : unit -> bool

Get the debugging flag of triggers.

val debug_interpretation : unit -> bool

Get the debugging flag for interpretation generatation.

Additional getters

Case-split options

val case_split_policy : unit -> AltErgoLib.Util.case_split_policy

Value specifying the case-split policy.

Possible values are :

after-theory-assume

before-matching

after-matching

Default to after-theory-assume

val enable_adts_cs : unit -> bool

true if case-split for Algebraic Datatypes theory is enabled.

Default to false

val max_split : unit -> AltErgoLib.Numbers.Q.t

Value specifying the maximum size of case-split.

Default to 1_000_000

Context options

val replay : unit -> bool

true if replay session will be saved in file_name.agr.

Default to false

val replay_all_used_context : unit -> bool

true if replay with all axioms and predicates saved in .used files of the current directory is enabled.

Default to false

val replay_used_context : unit -> bool

true if replay with axioms and predicates saved in .used file is enabled.

Default to false

val save_used_context : unit -> bool

true if used axioms and predicates will be saved in a .used file. This option implies unsat_core is set to true.

Default to false

Execution options

val answers_with_locs : unit -> bool

true if the locations of goals is shown when printing solver's answers.

Default to true

val frontend : unit -> string

Value of the currently selected parsing and typing frontend.

Default to legacy

val input_format : unit -> input_format

Value specifying the default input format. Useful when the extension does not allow to automatically select a parser (eg. JS mode, GUI mode, ...). possible values are

native

smtlib2

why3

Default to Native

val infer_input_format : unit -> bool

true if Alt-Ergo infers automatically the input format according to the file extension. false if an input format is set with -i option

Default to true

val parse_only : unit -> bool

true if the program shall stop after parsing.

Default to false

val parsers : unit -> string list

List of registered parsers for Alt-Ergo.

Default to false

val preludes : unit -> string list

List of files that have be loaded as preludes.

Default to []

val type_only : unit -> bool

true if the program shall stop after typing.

Default to false

val type_smt2 : unit -> bool

true if the program shall stop after SMT2 typing.

Default to false

Internal options

val disable_weaks : unit -> bool

true if the GC is prevented from collecting hashconsed data structrures that are not reachable (useful for more determinism).

Default to false

val enable_assertions : unit -> bool

true if verification of some heavy invariants is enabled.

Default to false

Limit options

val age_bound : unit -> int

Value specifying the age limit bound.

Default to 50

val fm_cross_limit : unit -> AltErgoLib.Numbers.Q.t

Value specifying the limit above which Fourier-Motzkin variables elimination steps that may produce a number of inequalities that is greater than this limit are skipped. However, unit eliminations are always done.

Default to 10_000

val steps_bound : unit -> int

Value specifying the maximum number of steps.

Default to -1

val timelimit : unit -> float

Value specifying the time limit (not supported on Windows).

Default to 0.

val timelimit_interpretation : unit -> float

Value specifying the time limit for model generation (not supported on Windows).

Default to 1. (not supported on Windows)

val timelimit_per_goal : unit -> bool

Value specifying the given timelimit for each goal in case of multiple goals per file. In this case, time spent in preprocessing is separated from resolution time.

$Not relevant for GUI-mode.

Default to false

Output options

val model : unit -> bool

Experimental support for models on labeled terms.

Possible values are

None

Default

Complete : shows a complete model

All : shows all models

Which are used in the two setters below. This option answers true if the model is set to Default or Complete

Default to false

val complete_model : unit -> bool

true if the model is set to complete model

Default to false

val all_models : unit -> bool

true if the model is set to all models?

Default to false

val interpretation : unit -> int

Experimental support for counter-example generation.

Possible values are :

Unknown

Before instantiation

Before every decision or instantiation

A negative value (-1, -2, or -3) will disable interpretation display.

Note that max_split limitation will be ignored in model generation phase.

Default to 0

val output_format : unit -> output_format

Value specifying the default output format. possible values are

native

smtlib2

why3

.

Default to Native

val infer_output_format : unit -> bool

true if Alt-Ergo infers automatically the output format according to the the file extension or the input format if set.

Default to true

val unsat_core : unit -> bool

true if experimental support for unsat-cores is on.

Default to false

Profiling options

val profiling : unit -> bool

true if the profiling module is activated.

Use Ctrl-C to switch between different views and Ctrl + AltGr + \ to exit.

Default to false

val profiling_period : unit -> float

Value specifying the profiling module frequency.

Default to 0.

val cumulative_time_profiling : unit -> bool

true if the time spent in called functions is recorded in callers

Default to false

val profiling_plugin : unit -> string

Value specifying which module is used as a profiling plugin.

Default to false

val timers : unit -> bool

true if profiling is set to true (automatically enabled)

Default to false

val verbose : unit -> bool

true if the verbose mode is activated.

Default to false

Quantifier options

val greedy : unit -> bool

true if all available ground terms are used in instantiation.

Default to false

val instantiate_after_backjump : unit -> bool

true if a (normal) instantiation round is made after every backjump/backtrack.

Default to false

val max_multi_triggers_size : unit -> int

Value specifying the max number of terms allowed in multi-triggers.

Default to 4

val triggers_var : unit -> bool

true if variables are allowed as triggers.

Default to false

val nb_triggers : unit -> int

Value specifying the number of (multi)triggers.

Default to 2

val no_ematching : unit -> bool

true if matching modulo ground equalities is disabled.

Default to false

val no_user_triggers : unit -> bool

true if user triggers are ignored except for triggers of theories axioms

Default to false

val normalize_instances : unit -> bool

true if generated substitutions are normalised by matching w.r.t. the state of the theory.

This means that only terms that are greater (w.r.t. depth) than the initial terms of the problem are normalized.

Default to false

SAT options

val arith_matching : unit -> bool

true if (the weak form of) matching modulo linear arithmetic is disabled.

Default to true

val no_backjumping : unit -> bool

true if backjumping mechanism in the functional SAT solver is disabled.

Default to false

val bottom_classes : unit -> bool

true if equivalence classes at each bottom of the SAT are shown.

Default to false

val sat_solver : unit -> AltErgoLib.Util.sat_solver

Value specifying which SAT solver is being used.

Possible values are:

CDCL-tableaux : CDCL SAT-solver with formulas filtering based on tableaux method

satML-Tableaux

satML-tableaux

CDCL : CDCL SAT-solver

satML

Tableaux : SAT-solver based on tableaux method

tableaux

tableaux-like

Tableaux-like

Tableaux-CDCL : Tableaux method assisted with a CDCL SAT-solver

tableaux-cdcl

tableaux-CDCL

Tableaux-cdcl

Default to CDCL-tableaux

val cdcl_tableaux_inst : unit -> bool

true if the use of a tableaux-like method for instantiations is enabled with the CDCL solver if satML is used.

Default to true

val cdcl_tableaux_th : unit -> bool

true if the use of a tableaux-like method for theories is enabled with the CDCL solver if satML is used.

Default to true

val cdcl_tableaux : unit -> bool

true if the use of a tableaux-like method for theories or instantiations is enabled with the CDCL solver if satML is used.

Default to true

val tableaux_cdcl : unit -> bool

true if the tableaux SAT-solver is used with CDCL assist.

Default to false

val minimal_bj : unit -> bool

true if minimal backjumping in satML CDCL solver is enabled

Default to true

val enable_restarts : unit -> bool

true if restarts are enabled for satML.

Default to false

val disable_flat_formulas_simplification : unit -> bool

true if facts simplification is disabled in satML's flat formulas.

Default to false

val no_sat_learning : unit -> bool

val sat_learning : unit -> bool

true if learning/caching of unit facts in the Default SAT is disabled. These facts are used to improve bcp.

Default to true (sat_learning is active)

val sat_plugin : unit -> string

Value specifying which SAT-solver is used.

Default to false

Term options

val disable_ites : unit -> bool

true if handling of ite(s) on terms in the backend is disabled.

Default to false

val inline_lets : unit -> bool

true if substitution of variables bounds by Let is enabled. The default behavior is to only substitute variables that are bound to a constant, or that appear at most once.

Default to false

val rewriting : unit -> bool

true if rewriting is used instead of axiomatic approach.

Default to false

val term_like_pp : unit -> bool

true if semantic values shall be output as terms.

Default to true

Theory options

val disable_adts : unit -> bool

true if Algebraic Datatypes theory is disabled

Default to false

val inequalities_plugin : unit -> string

Value specifying which module is used to handle inequalities of linear arithmetic.

Default to false

val no_ac : unit -> bool

true if the AC (Associative and Commutative) theory is disabled for function symbols.

Default to false

val no_contracongru : unit -> bool

true if contracongru is disabled.

Default to false

val no_fm : unit -> bool

true if Fourier-Motzkin algorithm is disabled.

Default to false

val no_nla : unit -> bool

true if non-linear arithmetic reasoning (i.e. non-linear multplication, division and modulo on integers and rationals) is disabled. Non-linear multiplication remains AC.

Default to false

val no_tcp : unit -> bool

true if BCP modulo theories is deactivated.

Default to false

val no_backward : unit -> bool

true if backward reasoning step (starting from the goal) done in the default SAT solver before deciding is disabled.

Default to false

val no_decisions : unit -> bool

true if decisions at the SAT level are disabled.

Default to false

val no_theory : unit -> bool

true if theory reasoning is completely deactivated.

Default to false

val restricted : unit -> bool

true if the set of decision procedures (equality, arithmetic and AC) is restricted.

Default to false

val tighten_vars : unit -> bool

true if the best bounds for arithmetic variables is computed.

Default to false

val use_fpa : unit -> bool

true if support for floating-point arithmetic is enabled.

Default to false

val rules : unit -> int

Possible values are

0 : parsing

1 : typing

2 : sat

3 : cc

4 : arith

output rules used on stderr.

Default to -1

Files

val js_mode : unit -> bool

true if the JavaScript mode is activated

Default to false

val get_file : unit -> string

Value specifying the file given to Alt-Ergo

Default to ""

val get_session_file : unit -> string

Value specifying the session file (base_name.agr) handled by Alt-Ergo

Default to false

val get_used_context_file : unit -> string

Value specifying the base name of the file (with no extension)

Default to false

Functions that are immediately executed

val exec_thread_yield : unit -> unit

val exec_timeout : unit -> unit

Simple Timer module

module Time : sig ... end

Globals

val tool_req : int -> string -> unit

Displays the used rules

Monomorphisations

val (<>) : int -> int -> bool

val (=) : int -> int -> bool

val (<) : int -> int -> bool

val (>) : int -> int -> bool

val (<=) : int -> int -> bool

val (>=) : int -> int -> bool

val compare : int -> int -> int

val can_decide_on : string -> bool

val no_decisions_on__is_empty : unit -> bool

val match_extension : string -> input_format

Extra

val set_is_gui : bool -> unit

val get_is_gui : unit -> bool

val parse_cmdline_arguments : unit -> unit